refactor: split builtin factory and knowledge public api
