+++ src/platform/knowledge/index.ts
@@ -1,0 +1,2 @@
+export { KnowledgeManager } from "./knowledge-manager.js";
+export * from "./public-api.js";
