+++ src/tools/index.ts
@@ -5,14 +5,15 @@ export type { ToolExecutorDeps } from "./executor.js";
 export { ToolPermissionManager } from "./permission.js";
 export type { PermissionManagerDeps, PermissionRule } from "./permission.js";
 export { ConcurrencyController } from "./concurrency.js";
-export { createBuiltinTools } from "./builtin/index.js";
-export type { BuiltinToolDeps } from "./builtin/index.js";
+export { createBuiltinTools } from "./builtin/factory.js";
+export type { BuiltinToolDeps } from "./builtin/factory.js";
 export { SkillSearchTool } from "./query/SkillSearchTool/SkillSearchTool.js";
 export {
   GitHubReadTool,
   GitHubPrCreateTool,
-  McpListToolsTool,
-  McpCallToolTool,
+} from "./network/GitHubCliTool/GitHubCliTool.js";
+export { McpListToolsTool, McpCallToolTool } from "./network/McpStdioTool/McpStdioTool.js";
+export {
   ProcessSessionManager,
   ProcessSessionStartTool,
   ProcessSessionReadTool,
@@ -20,5 +21,5 @@ export {
   ProcessSessionStopTool,
   ProcessSessionListTool,
   defaultProcessSessionManager,
-} from "./builtin/index.js";
+} from "./system/ProcessSessionTool/ProcessSessionTool.js";
 export * from "./types.js";
